Continue to work around RIC package versioning issues.
(After the cherry release, the ric-dep repo started naming newer
service versions in the undocumented example_recipe_oran_{version}
yaml files.  But they sort of maintain the documented main
example_recipe.yaml file.  It's hard to know at this point what
is really bronze/cherry/dawn.)

But at least for the appmgr, it's unclear to me if the cherry
0.5.1 and dawn 0.5.3 versions are complete yet, and the default
overrides stick with 0.4.3 .  So we will, too.